fix: move to bun
