+++ .changeset/violet-starfishes-act.md
@@ -1,0 +1,5 @@
+---
+"hopfield": patch
+---
+
+**Fix:** moved project to Bun.
