+++ .github/actions/install-dependencies/action.yml
@@ -4,15 +4,9 @@ description: "Prepare repository and all dependencies"
 runs:
   using: "composite"
   steps:
-    - name: Set up pnpm
-      uses: pnpm/action-setup@v2
-
-    - name: Set up node
-      uses: actions/setup-node@v3
-      with:
-        cache: pnpm
-        node-version: 18
+    - name: Set up Bun
+      uses: oven-sh/setup-bun@v1
 
     - name: Install dependencies
       shell: bash
-      run: pnpm install --ignore-scripts
+      run: bun install --ignore-scripts
