+++ .github/workflows/on-pull-request.yml
@@ -13,21 +13,6 @@ jobs:
     uses: ./.github/workflows/verify.yml
     secrets: inherit
 
-  # bench:
-  #   name: Benchmark
-  #   runs-on: ubuntu-latest
-  #   timeout-minutes: 5
-
-  #   steps:
-  #     - name: Clone repository
-  #       uses: actions/checkout@v4
-
-  #     - name: Install dependencies
-  #       uses: ./.github/actions/install-dependencies
-
-  #     - name: Run benchmarks
-  #       run: pnpm bench
-
   size:
     name: Size
     runs-on: ubuntu-latest
