+++ .github/workflows/on-push-to-main.yml
@@ -2,6 +2,7 @@ name: Main
 on:
   push:
     branches: [main]
+  workflow_dispatch:
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -13,41 +14,8 @@ jobs:
     uses: ./.github/workflows/verify.yml
     secrets: inherit
 
-  canary:
-    name: Release canary
-    needs: verify
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-
-    steps:
-      - name: Clone repository
-        uses: actions/checkout@v4
-
-      - name: Install dependencies
-        uses: ./.github/actions/install-dependencies
-
-      - name: Setup .npmrc file
-        uses: actions/setup-node@v3
-        with:
-          registry-url: 'https://registry.npmjs.org'
-
-      - name: Set version
-        run: |
-          npm --no-git-tag-version version minor
-          npm --no-git-tag-version version $(npm pkg get version | sed 's/"//g')-canary.$(date +'%Y%m%dT%H%M%S')
-          pnpm bun scripts/updateVersion.ts
-
-      - name: Build
-        run: pnpm build
-
-      - name: Publish to npm
-        run: npm publish --tag canary
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
-
   changesets:
     name: Changesets
-    needs: verify
     runs-on: ubuntu-latest
     timeout-minutes: 5
 
@@ -61,13 +29,32 @@ jobs:
       - name: Install dependencies
         uses: ./.github/actions/install-dependencies
 
-      - name: Create version pull request or publish to npm
+      - name: Create Version Pull Request
         uses: changesets/action@v1
         with:
-          title: 'chore: version packages'
-          commit: 'chore: version packages'
-          publish: pnpm changeset:release
-          version: pnpm changeset:version
+          version: bun run changeset:version
+          commit: 'chore: version package'
+          title: 'chore: version package'
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+  release:
+    name: Release
+    needs: verify
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+
+    steps:
+      - name: Clone repository
+        uses: actions/checkout@v4
+
+      - name: Install dependencies
+        uses: ./.github/actions/install-dependencies
+
+      - name: Publish to NPM
+        uses: changesets/action@v1
+        with:
+          publish: bun run changeset:release
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
