--- pnpm-workspace.yaml
@@ -1,5 +1,0 @@
-prefer-workspace-packages: true
-packages:
-  - "."
-  - "docs"
-  - "examples/*"
