+++ scripts/concatDocs.ts
@@ -49,7 +49,7 @@ const concatenateMarkdownFiles = async (
   }
 };
 
-const inputFolder = `${process.env.INIT_CWD}/docs`;
-const outputFile = `${process.env.INIT_CWD}/examples/next-13/src/app/docs.ts`;
+const inputFolder = './docs';
+const outputFile = './examples/next-13/src/app/docs.ts';
 
 concatenateMarkdownFiles(inputFolder, outputFile);
