+++ scripts/prepublishOnly.ts
@@ -1,20 +1,10 @@
-import { readJsonSync, writeJsonSync } from 'fs-extra';
-import path from 'path';
+import { join } from 'node:path';
 
-// Generates a package.json to be published to NPM with only the necessary fields.
-const packageJsonPath = path.join(__dirname, '../package.json');
-const tmpPackageJson = readJsonSync(packageJsonPath);
+const packageJsonPath = join(import.meta.dir, '../src/package.json');
+const packageJson = await Bun.file(packageJsonPath).json();
 
-writeJsonSync(`${packageJsonPath}.tmp`, tmpPackageJson, { spaces: 2 });
+// NOTE: We explicitly don't want to publish the type field.
+// We create a separate package.json for `dist/cjs` and `dist/esm` that has the type field.
+delete packageJson.type;
 
-const {
-  'simple-git-hooks': _sgh,
-  devDependencies: _dD,
-  packageManager: _pM,
-  pnpm: _p,
-  scripts: _s,
-  // NOTE: We explicitly don't want to publish the type field. We create a separate package.json for `dist/cjs` and `dist/esm` that has the type field.
-  type: _t,
-  ...rest
-} = tmpPackageJson;
-writeJsonSync(packageJsonPath, rest, { spaces: 2 });
+Bun.write(packageJsonPath, JSON.stringify(packageJson, null, 2));
