+++ test/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "test",
+  "private": true,
+  "type": "module"
+}
